requirements.txt
prometheus-client==0.14.1
PyYAML==6.0
